+++ src/FSharp.Literate/Contexts.fs
@@ -7,7 +7,7 @@ type CompilerContext =
   { // An instance of the F# code formatting agent
     FormatAgent : CodeFormatAgent 
     // F# interactive evaluator
-    Evaluator : FsiEvaluator option
+    Evaluator : IFsiEvaluator option
     // Command line options for the F# compiler
     CompilerOptions : string option
     // Defined symbols for the F# compiler